.gcloud/functions/load_data_into_bigquery/requirements.txt
google-cloud-bigquery==3.38.0
google-cloud-storage==3.5.0
